+++ DuckEngine/DuckEngine/include/DuckEditor/Layers/InspectorLayer.h
@@ -12,7 +12,7 @@ class InspectorLayer : public Layer
 {
 public:
 	InspectorLayer();
-	virtual void Update(f64 dt) override;
+	virtual void Update(f32 dt) override;
 
 	static LayerType GetType() { return LayerType::Inspector; }
 	virtual LayerType GetTypeDynamic() override { return GetType(); }
