+++ DuckEngine/DuckEngine/src/DuckEditor/HelperFunctions.cpp
@@ -4,12 +4,12 @@
 
 fm::vec2 IMToFM(ImVec2 vec)
 {
-	return {static_cast<double>(vec.x), static_cast<double>(vec.y)};
+	return {vec.x, vec.y};
 }
 
 ImVec2 FMToIM(fm::vec2 vec)
 {
-	return {static_cast<float>(vec.x), static_cast<float>(vec.y)};
+	return {vec.x, vec.y};
 }
 
 #endif
